+++ lib/modules/theme/theme_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: theme
-// Updated at: 2026-06-26 23:26:05
-// Commit: feat(theme): add async theme stream provider
+// Updated at: 2026-06-28 13:59:36
+// Commit: chore(theme): update dependencies and cleanup unused imports in theme
 
 class themeComponentState {
-  final String commitRef = '2390';
-  final String lastMessage = 'feat(theme): add async theme stream provider';
-  final DateTime timestamp = DateTime.parse('2026-06-26T23:26:05+05:00');
+  final String commitRef = '2392';
+  final String lastMessage = 'chore(theme): update dependencies and cleanup unused imports in theme';
+  final DateTime timestamp = DateTime.parse('2026-06-28T13:59:36+05:00');
 }